+++ web-test-runner.config.mjs
@@ -1,6 +1,7 @@
 import { esbuildPlugin } from '@web/dev-server-esbuild';
 import { playwrightLauncher } from '@web/test-runner-playwright';
 
+/** @type {import('@web/test-runner').TestRunnerConfig} */
 export default {
   nodeResolve: true,
   files: 'packages/**/*.test.ts',